p11
git-svn-id: svn://localhost@13 2093b545-fd6f-4e46-8590-e32825573395

+++ ref_gl.old/gl_model.h
@@ -119,7 +119,7 @@ typedef struct msurface_s
 
 	byte		styles[MAXLIGHTMAPS];
 	byte		*samples;		// [numstyles*surfsize]
-	float		cached_light[MAXLIGHTMAPS];	// values currently used in lightmap
+	byte		cached_light[MAXLIGHTMAPS];	// values currently used in lightmap
 
 	int			lightmaptexturenum;
 } msurface_t;
